training_set/requirements.txt
absl-py==0.1.13
alabaster==0.7.9
anaconda-clean==1.0
anaconda-client==1.5.1
anaconda-navigator==1.3.1
appdirs==1.4.0
argcomplete==1.0.0
astor==0.6.2
astroid==1.4.7
astropy==1.2.1
Athena-DL==0.6
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
beautifulsoup4==4.5.1
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.2
boto==2.42.0
boto3==1.7.6
botocore==1.10.6
Bottleneck==1.1.0
cachetools==2.0.1
certifi==2018.1.18
cffi==1.7.0
chardet==3.0.4
chest==0.2.3
click==3.3
cloudpickle==0.2.1
clyent==1.2.2
colorama==0.3.7
comtypes==1.1.2
conda==4.5.11
conda-build==2.0.2
configobj==5.0.6
contextlib2==0.5.3
cryptography==1.5
cycler==0.10.0
Cython==0.24.1
cytoolz==0.8.0
dask==0.17.2
datashape==0.5.2
decorator==4.0.10
dill==0.2.5
docutils==0.12
dynd===c328ab7
ecdsa==0.13
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.6
filterpy==0.1.5
Flask==0.11.1
Flask-Cors==2.1.2
Flask-Jsonpify==1.5.0
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gast==0.2.0
gevent==1.1.2
google-api-core==0.1.4
google-auth==1.3.0
google-auth-oauthlib==0.2.0
google-cloud==0.32.0
google-cloud-bigquery==0.29.0
google-cloud-bigquery-datatransfer==0.1.0
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.0
google-cloud-core==0.28.0
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.0
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.0
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.0
google-cloud-runtimeconfig==0.28.0
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.0
google-cloud-videointelligence==1.0.0
google-cloud-vision==0.29.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
graphviz==0.8
greenlet==0.4.10
grpc-google-iam-v1==0.11.4
grpcio==1.9.1
h5py==2.7.1
HeapDict==1.0.0
html5lib==0.9999999
httplib2==0.10.3
idna==2.6
imagesize==0.7.1
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.5.1
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
jupyter-dashboards==0.7.0
Keras==2.1.5
lazy-object-proxy==1.2.1
llvmlite==0.13.0
locket==0.2.0
lockfile==0.12.2
luigi==2.7.5
lxml==3.6.4
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==1.5.3
menuinst==1.4.1
mistune==0.7.3
mock==2.0.0
mpld3==0.3
mpmath==0.19
multipledispatch==0.4.8
mxnet==0.11.1b20170928
nb-anacondacloud==1.2.0
nb-conda==2.0.0
nb-conda-kernels==2.0.0
nbconvert==4.2.0
nbformat==4.4.0
nbpresent==3.0.2
networkx==1.11
nltk==3.2.1
nose==1.3.7
notebook==4.2.3
numba==0.28.1
numexpr==2.6.1
numpy==1.14.2
oauth2client==3.0.0
oauthlib==2.0.6
odo==0.5.0
olefile==0.44
openpyxl==2.3.2
packaging==16.8
pandas==0.22.0
pandas-datareader==0.6.0
pandasql==0.7.3
paramiko==1.15.2
partd==0.3.6
path.py==0.0.0
pathlib2==2.1.0
patsy==0.4.1
pbr==1.10.0
pep8==1.7.0
pickleshare==0.7.4
Pillow==4.3.0
pkginfo==1.3.2
plotly==2.1.0
ply==3.8
prompt-toolkit==1.0.3
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.5.1
psutil==5.4.3
py==1.4.31
py4j==0.10.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAthena==1.2.3
pycosat==0.6.3
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0
pydot==1.0.29
pyflakes==1.3.0
Pygments==2.1.3
pykalman==0.9.5
pylint==1.5.4
pyodbc==3.0.10
pyOpenSSL==16.2.0
pyparsing==2.1.10
pyspark==2.3.0
pytest==2.9.2
python-daemon==2.1.2
python-dateutil==2.5.3
pytz==2017.3
pywin32==220
PyYAML==3.12
pyzmq==15.4.0
QtAwesome==0.3.3
qtconsole==4.2.1
QtPy==1.1.2
regex
requests==2.18.4
requests-file==1.4.3
requests-ftp==0.3.1
requests-oauthlib==0.8.0
retrying==1.3.3
rope-py3k==0.9.4.post1
rsa==3.4.2
ruamel-yaml===-VERSION
s3transfer==0.1.13
scikit-image==0.12.3
scikit-learn==0.18.2
scipy==0.18.1
seaborn==0.9.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
sphinx==1.4.6
spyder==3.0.0
SQLAlchemy==1.0.13
statsmodels==0.6.1
sympy==1.0
tables==3.4.2
tenacity==4.10.0
tensorboard==1.7.0
tensorflow==1.7.0
tensorflow-gpu==1.2.0
tensorflowjs==0.1.1
tensorly==0.2.0
termcolor==1.1.0
-e git://github.com/Theano/Theano.git@fa9a870dff1b5973bebc56d9803357dad06e9e1c#egg=Theano
toolz==0.8.0
tornado==4.4.1
traitlets==4.3.0
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==1.2.6
win-unicode-console==0.5
WMI==1.4.9
wrapt==1.10.6
xlrd==1.0.0
XlsxWriter==0.9.3
xlwings==0.10.0
xlwt==1.1.2
